+++ src/App/App.module.scss
@@ -1,0 +1,3 @@
+.app {
+  text-align: center;
+}
