+++ src/App/App.tsx
@@ -1,0 +1,13 @@
+import styles from "./App.module.scss";
+
+function App() {
+  return (
+    <div className={styles.app}>
+      <header className={styles.header}>
+        <p>Github Client</p>
+      </header>
+    </div>
+  );
+}
+
+export default App;
